Ensure application order is respected in cascade
Dropping results from the df might lead to df["application"].unique()
returning a different list of applications. The order of applications in
this list determines the order in which applications are plotted.

In order to respect the order of applications in the original DataFrame,
and thereby ensure that cascade and navchart plots produced from the
same projection use the same legend, we need to ensure that we do not
change the application order.
